+++ src/backend/base/langflow/components/data/__init__.py
@@ -3,6 +3,7 @@ from .csv_to_data import CSVToDataComponent
 from .directory import DirectoryComponent
 from .file import FileComponent
 from .json_to_data import JSONToDataComponent
+from .mcp_component import MCPToolsComponent
 from .rss import RSSReaderComponent
 from .sql_executor import SQLComponent
 from .url import URLComponent
@@ -15,6 +16,7 @@ __all__ = [
     "DirectoryComponent",
     "FileComponent",
     "JSONToDataComponent",
+    "MCPToolsComponent",
     "RSSReaderComponent",
     "SQLComponent",
     "URLComponent",
