+++ src/backend/base/langflow/components/helpers/__init__.py
@@ -1,3 +1,4 @@
+from .calculator_core import CalculatorComponent
 from .create_list import CreateListComponent
 from .current_date import CurrentDateComponent
 from .id_generator import IDGeneratorComponent
@@ -8,6 +9,7 @@ from .structured_output import StructuredOutputComponent
 
 __all__ = [
     "BatchRunComponent",
+    "CalculatorComponent",
     "CreateListComponent",
     "CurrentDateComponent",
     "IDGeneratorComponent",
