+++ src/backend/base/langflow/components/processing/__init__.py
@@ -12,6 +12,7 @@ from .message_to_data import MessageToDataComponent
 from .parse_data import ParseDataComponent
 from .parse_json_data import ParseJSONDataComponent
 from .parser import ParserComponent
+from .python_repl_core import PythonREPLComponent
 from .regex import RegexExtractorComponent
 from .select_data import SelectDataComponent
 from .split_text import SplitTextComponent
@@ -33,6 +34,7 @@ __all__ = [
     "ParseDataFrameComponent",
     "ParseJSONDataComponent",
     "ParserComponent",
+    "PythonREPLComponent",
     "RegexExtractorComponent",
     "SelectDataComponent",
     "SplitTextComponent",
