+++ src/backend/base/langflow/components/tools/__init__.py
@@ -3,13 +3,10 @@ import warnings
 from langchain_core._api.deprecation import LangChainDeprecationWarning
 
 from .calculator import CalculatorToolComponent
-from .calculator_core import CalculatorComponent
 from .google_search_api import GoogleSearchAPIComponent
 from .google_serper_api import GoogleSerperAPIComponent
-from .mcp_component import MCPToolsComponent
 from .python_code_structured_tool import PythonCodeStructuredTool
 from .python_repl import PythonREPLToolComponent
-from .python_repl_core import PythonREPLComponent
 from .search_api import SearchAPIComponent
 from .searxng import SearXNGToolComponent
 from .serp import SerpComponent
@@ -26,16 +23,13 @@ with warnings.catch_warnings():
 __all__ = [
     "AstraDBCQLToolComponent",
     "AstraDBToolComponent",
-    "CalculatorComponent",
     "CalculatorToolComponent",
     "DuckDuckGoSearchComponent",
     "ExaSearchToolkit",
     "GleanSearchAPIComponent",
     "GoogleSearchAPIComponent",
     "GoogleSerperAPIComponent",
-    "MCPToolsComponent",
     "PythonCodeStructuredTool",
-    "PythonREPLComponent",
     "PythonREPLToolComponent",
     "SearXNGToolComponent",
     "SearchAPIComponent",
