+++ src/backend/tests/integration/components/mcp/test_mcp_component.py
@@ -2,7 +2,7 @@ from tests.integration.utils import run_single_component
 
 
 async def test_mcp_component():
-    from langflow.components.tools.mcp_component import MCPToolsComponent
+    from langflow.components.data.mcp_component import MCPToolsComponent
 
     inputs = {}
     await run_single_component(
