+++ src/backend/tests/unit/components/tools/test_calculator.py
@@ -1,5 +1,5 @@
 import pytest
-from langflow.components.tools.calculator_core import CalculatorComponent
+from langflow.components.helpers.calculator_core import CalculatorComponent
 
 from tests.base import ComponentTestBaseWithoutClient
