+++ src/backend/tests/unit/components/tools/test_mcp_component.py
@@ -2,7 +2,7 @@ import asyncio
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
-from langflow.components.tools.mcp_component import MCPSseClient, MCPStdioClient, MCPToolsComponent
+from langflow.components.data.mcp_component import MCPSseClient, MCPStdioClient, MCPToolsComponent
 
 from tests.base import ComponentTestBaseWithoutClient, VersionComponentMapping
